+++ demo/index.ts
@@ -1,44 +1,4 @@
-import {bootstrap} from '@angular/platform-browser-dynamic';
-import {Component} from '@angular/core';
-import {NgClass} from '@angular/common';
+import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
+import { NgFileUploadDemo } from './ng2-file-upload-demo.module';
 
-import {FileUploadSectionComponent} from './components/file-upload-section';
-
-let gettingStarted = require('./getting-started.md');
-
-@Component({
-  selector: 'app',
-  template: `
-  <main class="bd-pageheader">
-    <div class="container">
-      <h1>ng2-file-upload</h1>
-      <p>The Angular2 File Upload directives</p>
-      <a class="btn btn-primary" href="https://github.com/valor-software/ng2-file-upload">View on GitHub</a>
-      <div class="row">
-        <div class="col-lg-1"><iframe src="https://ghbtns.com/github-btn.html?user=valor-software&repo=ng2-file-upload&type=star&count=true" frameborder="0" scrolling="0" width="170px" height="20px"></iframe></div>
-        <div class="col-lg-1"><iframe src="https://ghbtns.com/github-btn.html?user=valor-software&repo=ng2-file-upload&type=fork&count=true" frameborder="0" scrolling="0" width="170px" height="20px"></iframe></div>
-      </div>
-    </div>
-  </main>
-
-  <div class="container">
-    <section id="getting-started">${gettingStarted}</section>
-
-    <file-upload-section class="col-md-12"></file-upload-section>
-  </div>
-
-  <footer class="footer">
-    <div class="container">
-      <p class="text-muted text-center"><a href="https://github.com/valor-software/ng2-file-upload">ng2-file-upload</a> is maintained by <a href="https://github.com/valor-software">valor-software</a>.</p>
-    </div>
-  </footer>
-  `,
-  directives: [
-    NgClass,
-    FileUploadSectionComponent
-  ]
-})
-export class DemoComponent {
-}
-
-bootstrap(DemoComponent);
+platformBrowserDynamic().bootstrapModule(NgFileUploadDemo);
