+++ package.json
@@ -48,15 +48,16 @@
   "homepage": "https://github.com/valor-software/ng2-file-upload#readme",
   "dependencies": {},
   "peerDependencies": {
-    "@angular/common": "^2.0.0-rc.1",
-    "@angular/core": "^2.0.0-rc.1"
+    "@angular/common": "2.0.0-rc.6",
+    "@angular/core": "2.0.0-rc.6"
   },
   "devDependencies": {
-    "@angular/common": "^2.0.0-rc.1",
-    "@angular/compiler": "^2.0.0-rc.1",
-    "@angular/core": "^2.0.0-rc.1",
-    "@angular/platform-browser": "^2.0.0-rc.1",
-    "@angular/platform-browser-dynamic": "^2.0.0-rc.1",
+    "@angular/common": "2.0.0-rc.6",
+    "@angular/compiler": "2.0.0-rc.6",
+    "@angular/core": "2.0.0-rc.6",
+    "@angular/platform-browser": "2.0.0-rc.6",
+    "@angular/platform-browser-dynamic": "2.0.0-rc.6",
+    "@angular/forms": "2.0.0-rc.6",
     "async": "1.5.2",
     "bootstrap": "3.3.6",
     "codecov": "1.0.1",
@@ -92,7 +93,7 @@
     "lite-server": "2.2.0",
     "markdown-loader": "0.1.7",
     "marked": "0.3.5",
-    "ng2-bootstrap": "1.0.17",
+    "ng2-bootstrap": "1.1.1",
     "phantomjs-polyfill": "0.0.2",
     "phantomjs-prebuilt": "2.1.7",
     "pre-commit": "1.1.3",
@@ -101,7 +102,7 @@
     "raw-loader": "0.5.1",
     "reflect-metadata": "0.1.2",
     "require-dir": "0.3.0",
-    "rxjs": "5.0.0-beta.6",
+    "rxjs": "5.0.0-beta.11",
     "source-map-loader": "0.1.5",
     "systemjs-builder": "0.15.19",
     "ts-loader": "0.8.2",
@@ -110,7 +111,7 @@
     "typings": "0.8.1",
     "webpack": "1.13.1",
     "webpack-dev-server": "1.14.1",
-    "zone.js": "0.6.12"
+    "zone.js": "0.6.17"
   },
   "contributors": [
     {
